requirements.txt
pandas>=1.3.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
requests>=2.26.0
sqlalchemy>=1.4.0